refactor: make agent workflow page server component
